PlantDisease/requirements.txt
flask
flask-cors
torch
torchvision
transformers
pillow
python-dotenv
google-generativeai
evaluate
accelerate
scikit-learn
numpy
pandas
